pyproject.toml
[tool.poetry]
name = "fastapi-terraform-gke-example"
version = "0.1.0"
description = "Basic FastAPI setup using elements of DDD"
authors = ["Alexander Kupreev <agkupreev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
uvicorn = {extras = ["standard"],version = "<0.14.0,>=0.12.0"}
fastapi = "^0.63.0"
gunicorn = "^20.0.4"
pydantic = "^1.4"
sqlalchemy = "^1.3.16"
psycopg2-binary = "^2.8.6"
alembic = "^1.5.7"
requests = "^2.25.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
tenacity = "^7.0.0"
python-multipart = "^0.0.5"
emails = "^0.6.0"
jinja2 = "^2.11.3"
email-validator = "^1.1.2"
python-jose = {extras = ["cryptography"], version = "^3.2.0"}
sh = "^1.14.1"
google-cloud-pubsub = "^2.5.0"
google-cloud-logging = "^2.4.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
mypy = "^0.812"
black = "^20.8b0"
isort = "^5.7.0"
flake8 = "^3.9.0"
pylint = "^2.7.1"
sqlalchemy-stubs = "^0.4"
pytest-cov = "^2.11.1"
pytest-dotenv = "^0.5.2"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
